update filepathfor macos version
use "/" instead of "\\"

+++ src/getdata.jl
@@ -1,3 +1,4 @@
+# module GetData
 using HTTP, JSON, DataFrames
 using ProgressMeter
 
@@ -11,13 +12,20 @@ using ProgressMeter
 #save the data to a json file separate by the filename 
 #filename = "ECS" * string(i) * ".json"
 
+println("TO DO
+Test files with the same name, after a couple of months, if they are the same, we can ignore running it\n 
+and only run the new one.")
+
+
 include("key.jl")
 
 function get_data(cate::String, page::Int64
     ; token::String=token, mode::Int64=0)
     if mode == 0
         res = HTTP.request("GET",
-            "https://buildingtransparency.org/api/$cate/?page_number=$page", #materials?page_number=2",
+            "https://buildingtransparency.org/api/$cate?page_number=$page", #materials?page_number=2",
+            # "https://buildingtransparency.org/api/$cate?name__like=cement&page_number=$page", #materials?page_number=2",
+
             ["Authorization" => "Bearer " * "$token"]
         )
     else
@@ -28,22 +36,11 @@ function get_data(cate::String, page::Int64
     return res
 end
 
-
-#test run to get the total number of pages first
-res = get_data("materials", 1)
-response_text = String(res.body)
-msg = JSON.parse(response_text)
-df = DataFrame(msg)
-##########
-# 06 Aug 
-# X-Total-Count: 147130
-# X-Total-Pages: 1472
-
-function scrapeit(;all::Bool=false, total_pages::Int64=5)
+function scrapeit(;all::Bool=false, total_pages::Int64=5, path = "rawdata/")
     cate = "materials"
     page = 1 #start from page 1
 
-    filepath = joinpath(@__DIR__, "rawdata/")
+    filepath = joinpath(@__DIR__, path)
 
     if all 
         total_pages = parse(Int64, (res["X-Total-Pages"]))
@@ -56,45 +53,68 @@ function scrapeit(;all::Bool=false, total_pages::Int64=5)
 
     #initialize variables
     p = Progress(total_pages)
-    update!(p,1)
+    ProgressMeter.update!(p,1)
     jj = Threads.Atomic{Int}(0)
     l = Threads.SpinLock()
 
     Threads.@threads for page = 1:total_pages
-
-        res = get_data(cate, page)
-        # if page == 1 & total_pages != 10 #first run
-        #     total_pages = parse(Int64, (res["X-Total-Pages"]))
-        #     println("Scraping process started")
-        #     println("There are :", total_pages, " pages")
-        # end
-
-        response_text = String(res.body)
-
+            # if page == 1 & total_pages != 10 #first run
+            #     total_pages = parse(Int64, (res["X-Total-Pages"]))
+            #     println("Scraping process started")
+            #     println("There are :", total_pages, " pages")
+            # end
         #get number to 0000 format for filename
         spg = "0"^(4 - length(string(page))) * string(page)
         filename = "pg" * string(spg) * ".json"
 
-        msg = JSON.parse(response_text)
-        #write a to a json file
-        # println(filepath * filename)
-        open(filepath * filename, "w") do f
-            JSON.print(f, msg)
+        if isfile(filepath * filename)
+            println("Found file ", filename, "-> Skipped")
+            Threads.atomic_add!(jj, 1)
+            Threads.lock(l)
+            ProgressMeter.update!(p, jj[])
+            Threads.unlock(l) 
+        else
+            res = get_data(cate, page)
+            response_text = String(res.body)
+            msg = JSON.parse(response_text)
+            #write a to a json file
+            # println(filepath * filename)
+            open(filepath * filename, "w") do f
+                JSON.print(f, msg)
+            end
+            # println("Page ", page, " done")
+            res = nothing #clear the memory
+            # if page == 2 
+            #     println("Test Run Ended")
+            #     break
+            # end
+            Threads.atomic_add!(jj, 1)
+            Threads.lock(l)
+            ProgressMeter.update!(p, jj[])
+            Threads.unlock(l) 
         end
-        # println("Page ", page, " done")
-        res = nothing #clear the memory
-        # if page == 2 
-        #     println("Test Run Ended")
-        #     break
-        # end
-        Threads.atomic_add!(jj, 1)
-        Threads.lock(l)
-        update!(p, jj[])
-        Threads.unlock(l)  
+ 
     end
     println("Scraping process ended")
     println("#"^20)
 end
 
-# @time scrapeit()
-@time scrapeit(all=true)
+
+#test run to get the total number of pages first
+res = get_data("materials", 1)
+n = Dict(res.headers)["X-Total-Count"]
+fullpath = joinpath(@__DIR__, "rawdata_$n/")
+println("Current fullpath: $fullpath")
+try
+    mkdir(fullpath)
+catch
+    println("Path:$fullpath is already exist")
+    println("Please proceed with care")
+end
+response_text = String(res.body)
+msg = JSON.parse(response_text)
+# df = DataFrame(msg)
+##########
+# 06 Aug 
+# X-Total-Count: 147130
+# X-Total-Pages: 1472
